requirements.txt
torch>=2.0.0
transformers>=4.40.0
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0
